FF: Fixes so that Py3 won't fall down with bad expInfo tokens
Ths fixes an issue with expInfo = "{'session':001}" where 001 needs
to be converted to a string first (only in Python3)
